crawler-service/requirements.txt
sqlalchemy>=2.0.25,<3
scrapy==2.11.0
scrapy-playwright==0.0.4
itemloaders==1.1.0
parsel==1.8.1
w3lib==2.1.2
lxml==5.1.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
